+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-m5kdev
 
+## 0.19.1
+
+### Patch Changes
+
+- table empty
+
 ## 0.19.0
 
 ### Minor Changes