+++ packages/frontend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @m5kdev/frontend
 
+## 0.19.1
+
+### Patch Changes
+
+- table empty
+- Updated dependencies
+  - @m5kdev/commons@0.19.1
+
 ## 0.19.0
 
 ### Minor Changes